[enh] Refactor finiteness_checker (#2669)
* Add avx2 support

* clang-formatting

* __mm256_set1_epi64 -> __mm256_set1_epi64x

* Update finiteness_checker.cpp

* clang-formatting

* switch avx512 avx2 ordering

* clang-formatting

* remove print statements

* isolate in macro ifs

* clang-formatting

* add cpu and header

* centralize to header file

* readd base case

* removed valuesarenotfinite?

* hanging chad

* attempts at separating avx512

* forgotten saves

* current status moving things to header|

* mangled to see if we can get avx512 to work again

* fix service_dispatch.h

* remove accidental addition

* arm also uses __DAAL_KERNEL_MIN

* remove whitespace

* first success

* refactor

* readd due to linking issue

* clang-formatting

* remove print statements

* this will break, but needed to store progress'

* swap to avx2 intrinsics

* swap to avx2 intrinsics

* r -> t

* remove commments, clean up

* fix typing problem in avx512

* move simd specifics to impl.i files

* clang-format and compilability

* forgotten comments

* templating

* Revert "templating"

This reverts commit 1c9ee4356c8d76294ea4a4e8fdc79d6827f6f4dc.

* remove initial duplication via templating sums

* will come back to this point later

* clang-formatting and forgotten  file

* add copyrights

* end space

* changes to match review

* last reviewer changes

* Update finiteness_checker_cpu.cpp

* another try at fixing bazel build issues

* Update BUILD

* formatting

* forgotten AVX naming leading to infinte loop

* add comments

* small docstring updates

* more documentation

* oops
