+++ cpp/daal/src/externals/service_dispatch.h
@@ -29,7 +29,8 @@
 
 #if defined(TARGET_X86_64)
     #define DAAL_DISPATCH_FUNCTION_BY_CPU(func, ...)                                                                    \
-        switch (static_cast<daal::CpuType>(daal::services::Environment::getInstance()->getCpuId()))                     \
+        switch (__DAAL_KERNEL_MIN(DAAL_KERNEL_BUILD_MAX_INSTRUCTION_SET_ID,                                             \
+                                  static_cast<daal::CpuType>(daal::services::Environment::getInstance()->getCpuId())))  \
         {                                                                                                               \
             DAAL_KERNEL_SSE42_ONLY_CODE(case daal::CpuType::sse42 : func(daal::CpuType::sse42, __VA_ARGS__); break;)    \
             DAAL_KERNEL_AVX2_ONLY_CODE(case daal::CpuType::avx2 : func(daal::CpuType::avx2, __VA_ARGS__); break;)       \
@@ -37,17 +38,19 @@
             DAAL_EXPAND(default : func(daal::CpuType::sse2, __VA_ARGS__); break;)                                       \
         }
 
-    #define DAAL_DISPATCH_FUNCTION_BY_CPU_SAFE(func, ...)                                                                    \
-        services::Status st;                                                                                                 \
-        int cpuid = daal::sse2;                                                                                              \
-        DAAL_SAFE_CPU_CALL((cpuid = daal::services::Environment::getInstance()->getCpuId()), (cpuid = daal::sse2))           \
-        switch (static_cast<daal::CpuType>(cpuid))                                                                           \
-        {                                                                                                                    \
-            DAAL_KERNEL_SSE42_ONLY_CODE(case daal::CpuType::sse42 : st = func(daal::CpuType::sse42, __VA_ARGS__); break;)    \
-            DAAL_KERNEL_AVX2_ONLY_CODE(case daal::CpuType::avx2 : st = func(daal::CpuType::avx2, __VA_ARGS__); break;)       \
-            DAAL_KERNEL_AVX512_ONLY_CODE(case daal::CpuType::avx512 : st = func(daal::CpuType::avx512, __VA_ARGS__); break;) \
-            DAAL_EXPAND(default : st = func(daal::CpuType::sse2, __VA_ARGS__); break;)                                       \
-        }                                                                                                                    \
+    #define DAAL_DISPATCH_FUNCTION_BY_CPU_SAFE(func, ...)                                                                                  \
+        services::Status st;                                                                                                               \
+        int cpuid = daal::sse2;                                                                                                            \
+        DAAL_SAFE_CPU_CALL(                                                                                                                \
+            (cpuid = __DAAL_KERNEL_MIN(DAAL_KERNEL_BUILD_MAX_INSTRUCTION_SET_ID, daal::services::Environment::getInstance()->getCpuId())), \
+            (cpuid = daal::sse2))                                                                                                          \
+        switch (static_cast<daal::CpuType>(cpuid))                                                                                         \
+        {                                                                                                                                  \
+            DAAL_KERNEL_SSE42_ONLY_CODE(case daal::CpuType::sse42 : st = func(daal::CpuType::sse42, __VA_ARGS__); break;)                  \
+            DAAL_KERNEL_AVX2_ONLY_CODE(case daal::CpuType::avx2 : st = func(daal::CpuType::avx2, __VA_ARGS__); break;)                     \
+            DAAL_KERNEL_AVX512_ONLY_CODE(case daal::CpuType::avx512 : st = func(daal::CpuType::avx512, __VA_ARGS__); break;)               \
+            DAAL_EXPAND(default : st = func(daal::CpuType::sse2, __VA_ARGS__); break;)                                                     \
+        }                                                                                                                                  \
         services::throwIfPossible(st);
 #elif defined(TARGET_ARM)
     #define DAAL_DISPATCH_FUNCTION_BY_CPU(func, ...)                                                           \
